requirements.txt
asgiref==3.2.7
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
numpy==1.18.4
Pillow==7.1.2
PyQRCode==1.2.1
pytz==2019.3
sqlparse==0.3.1
